Get Posterior Accumulation Rates

Get Posterior Accumulation Rates

Usage

get_posterior_acc_rates(hamstr_fit, tau = 0, kern = c("U", "G", "BH"))

Arguments

hamstr_fit

the object returned from stan_hamstr.

tau

scale of the smoothing kernel in depth units. If tau > 0, accumulation rates are smoothed (filtered) before summary statistics are calculated, so that the accumulation rate at a given depth corresponds to the average rate over the depth interval tau. Default to 0.

kern

choice of smoothin kernal. U for uniform (moving average), G for Gaussian, BH for Berger and Heath (exponential). Defaults to U

Value

a dataframe/tibble with posterior ages for all iterations after warmup

Examples

if (FALSE) { # \dontrun{
fit <- hamstr(
  depth = MSB2K$depth,
  obs_age = MSB2K$age,
  obs_err = MSB2K$error)

get_posterior_acc_rates(fit)
} # }