Return a List of hamstr Control Arguments

Returns a list arguments to pass to hamstr_control which may be useful for debugging or development

Usage

hamstr_control(
  scale_R = TRUE,
  nu = 6,
  scale_shape = TRUE,
  smooth_s = FALSE,
  inflate_errors = FALSE,
  infl_sigma_sd = NULL,
  infl_shape_shape = 1,
  infl_shape_mean = 1
)

Arguments

scale_R

logical: Scale AR1 coefficient by delta_c (as in Bacon) or not. Defaults to TRUE.

nu

degrees of freedom for the Student-t distributed error model. Defaults to 6, which is equivalent to the default parametrisation of t.a=3, t.b=4 in Bacon 2.2. Set to a high number to approximate a Gaussian error model, (nu = 100 should do it).

scale_shape

scale the shape parameter according to the number of hierarchical levels, to control the total variance of the alpha innovations. This defaults to TRUE as of hamstr verion 0.5.

smooth_s

smooth the sedimentation rate used to calculate additional error from bioturbation by taking a mean across nearby sections

inflate_errors

logical: If set to TRUE, observation errors are inflated so that data are consistent with a gamma AR1 age-depth model. This is an experimental feature under active development. Defaults to FALSE.

infl_sigma_sd

hyperparameter: sets the standard deviation of the half-normal prior on the mean of the additional error terms. Defaults to 10 times the mean observation error in obs_err.

infl_shape_shape, infl_shape_mean

hyperparameters: parametrises the gamma prior on the shape of the distribution of the additional error terms. Default to 1, 1.

Value

a named list

Examples

hamstr_control()
#> $scale_R
#> [1] TRUE
#> 
#> $nu
#> [1] 6
#> 
#> $scale_shape
#> [1] TRUE
#> 
#> $smooth_s
#> [1] FALSE
#> 
#> $inflate_errors
#> [1] FALSE
#> 
#> $infl_sigma_sd
#> NULL
#> 
#> $infl_shape_shape
#> [1] 1
#> 
#> $infl_shape_mean
#> [1] 1
#>